3rd_Year/3rd_Year_Project/code/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-environ
django-cors-headers
whitenoise
gunicorn==20.1.0